+++ src/graysvr/CObjBase.h
@@ -2760,7 +2760,7 @@ public:
 	DIR_TYPE m_dirFace;			// facing this dir.
 	CGString m_sTitle;			// Special title such as "the guard" (replaces the normal skill title).
 	CPointMap m_ptHome;			// What is our "home" region. (towns and bounding of NPC's)
-	INT64 m_virtualGold;		// Virtual gold used by TOL clients
+	unsigned long long m_virtualGold;	// Virtual gold used by TOL clients
 
 	// Speech
 	FONT_TYPE m_fonttype;		// speech font to use (client send this to server, but it's not used)
